+++ HISTORY.md
@@ -1,5 +1,17 @@
 # ModClean History
 
+## 1.2.0 (4/28/2015)
+* Added several new patterns to `patterns.json`.
+* Added ability to delete empty directories.
+    * Added `ModClean()._removeEmpty()`.
+    * Added `removeEmptyDirs` option.
+    * Added `-d, --empty` flags to the CLI to toggle `removeEmptyDirs` on.
+* Added `node-dir` dependency.
+* Added `BENCHMARK.md` file to store a list of benchmarks.
+* Updated `.travis.yml` to include io.js
+* Updated and fixed `README.md`.
+* Updated tests.
+
 ## 1.1.2 (4/25/2015)
 * Fixed line endings for OS X and Linux systems